Sema: Deterministically order declarations in EmittedMemberRequest
Previously we did this in the SILVTableVisitor, since maintaining a
consistent order between translation units is important for correctness.

However, we also want the order in which the members themselves are
emitted to remain consistent regardless of the order in which members
were synthesized. This is in service of reproducible builds.
